+++ packages/create-lint/template/test-jest-vue/config.js
@@ -9,7 +9,9 @@ module.exports = (options) => {
     pkg: [
       'jest',
       'ts-jest',
+      'ts-node',
       'jest-transform-stub',
+      'jest-environment-jsdom',
       '@types/jest',
       '@vue/vue3-jest',
       '@vue/test-utils',