+++ packages/create-lint/template/test-vitest-react/config.js
@@ -1,18 +1,18 @@
 const { setPkg } = require('../../utils/index')
 
 // https://github.com/vitest-dev/vitest
-// https://github.com/vuejs/test-utils
+// https://github.com/testing-library/react-testing-library
 module.exports = (options) => {
   return {
     pkg: [
       'vitest',
       '@vitest/ui',
       '@vitest/coverage-v8',
-      '@vitejs/plugin-vue',
-      '@vue/test-utils',
+      '@vitejs/plugin-react',
+      '@testing-library/react',
       'happy-dom',
     ],
-    configFile: ['test-vitest-vue/vitest.config.ts'],
+    configFile: ['test-vitest-react/vitest.config.ts'],
     async afterInstall() {
       await setPkg(options.cwd, {
         scripts: {