+++ packages/create-lint/template/test-vitest-react/vitest.config.ts
@@ -1,10 +1,10 @@
-import Vue from '@vitejs/plugin-vue'
+import React from '@vitejs/plugin-react'
 import { defineConfig } from 'vitest/config'
 
 // https://cn.vitest.dev/config/
 // Can write vite.config.ts if it already exists
 export default defineConfig({
-  plugins: [Vue()],
+  plugins: [React()],
   test: {
     environment: 'happy-dom',
   },